requirements.txt
python-dotenv>=1.0.0
google-genai>=0.3.0
pymupdf>=1.24.0
streamlit>=1.35.0
